# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateProjectRequest": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "path": {
            "description": "absolute path of a new (empty or missing) directory",
            "format": "path",
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path",
          "name"
        ],
        "title": "CreateProjectRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OpenProjectRequest": {
        "additionalProperties": false,
        "properties": {
          "path": {
            "description": "absolute path of an existing project directory",
            "format": "path",
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path"
        ],
        "title": "OpenProjectRequest",
        "type": "object"
      },
...
      "ProjectInfo": {
        "additionalProperties": false,
        "properties": {
          "manifest": {
            "$ref": "#/components/schemas/ProjectManifest"
          },
          "path": {
            "format": "path",
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path",
          "manifest"
        ],
        "title": "ProjectInfo",
        "type": "object"
      },
      "ProjectManifest": {
        "additionalProperties": false,
        "properties": {
          "calculation_ids": {
            "items": {
              "type": "string"
            },
            "title": "Calculation Ids",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "dataset_ids": {
            "items": {
              "type": "string"
            },
            "title": "Dataset Ids",
            "type": "array"
          },
          "format_version": {
            "default": 1,
            "title": "Format Version",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "modified_at": {
            "format": "date-time",
            "title": "Modified At",
            "type": "string"
          },
          "name": {
            "default": "untitled project",
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "default": "",
            "title": "Notes",
            "type": "string"
          },
          "structure_ids": {
            "items": {
              "type": "string"
            },
            "title": "Structure Ids",
            "type": "array"
          },
          "view_settings": {
            "additionalProperties": true,
            "title": "View Settings",
            "type": "object"
          }
        },
        "title": "ProjectManifest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Atomscope API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/project/create": {
      "post": {
        "operationId": "create_project_api_project_create_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Project",
        "tags": [
          "project"
        ]
      }
    },
    "/api/project/open": {
      "post": {
        "operationId": "open_project_api_project_open_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Open Project",
        "tags": [
          "project"
        ]
      }
    },
...
  }
}